+++ test/cases/include-alias-resolve.alias/webpack.config.js
@@ -34,7 +34,7 @@ module.exports = {
         test: /\.pug$/,
         loader: 'pug-loader',
         options: {
-          method: 'compile',
+          method: 'render',
         },
       },
     ],